requirements.txt
discord.py[voice]
youtube_dl
PyNaCl
giphy_client
imgurpython
requests
praw
twilio
bs4
googletrans
selenium